test: Refactor code to fetch proofs from server
Refactor this code to fetch proofs from server so it can be used for
other data as well, specifically to also fetch blk files containing
block data.
